Forgot the small commits so basically. This is the 0.1 version
It does however not work. Because it makes a depth first search on the first link it finds and it does not do this concurrently with other recursive calls so it basically just goes in total depth with the first link it finds. Secondly i dont think that it even finds the correct links. So if you dont mind waiting an eternity to find out if this version actually works or not, this commit is for you ;)